requirements.txt
pandas
cryptography==39.0.2
pyarrow==11.0.0
azure-keyvault-secrets==4.7.0
azure-identity==1.12.0
